Select where to deploy based on repo secret variable
Update deploy action to use the ID of Scrapy Cloud project defined in
the repository secrets, instead of the default value. This is important
so we can use a different Scrapy Cloud project for testing if needed or
to allow user forks to be able to use the action to deploy in different
test environemnts.